Using integer for vertex ids rather than longs

+++ modules/core/src/main/java/edu/agh/iga/adi/giraph/core/IgaMessage.java
@@ -2,20 +2,20 @@ package edu.agh.iga.adi.giraph.core;
 
 public abstract class IgaMessage {
 
-  private long srcId;
+  private int srcId;
   private IgaOperation operation;
 
-  protected IgaMessage(long srcId, IgaOperation operation) {
+  protected IgaMessage(int srcId, IgaOperation operation) {
     this.srcId = srcId;
     this.operation = operation;
   }
 
-  public IgaMessage reattach(long srcId) {
+  public IgaMessage reattach(int srcId) {
     this.srcId = srcId;
     return this;
   }
 
-  public long getSrcId() {
+  public int getSrcId() {
     return srcId;
   }
