+++ modules/direction/src/main/java/edu/agh/iga/adi/giraph/direction/io/data/message/BackwardsSubstituteInterimMessageSerializer.java
@@ -20,14 +20,14 @@ final class BackwardsSubstituteInterimMessageSerializer implements MessageSerial
 
   @Override
   public void writeMessage(DataOutput dataOutput, BackwardsSubstituteInterimMessage message) throws IOException {
-    dataOutput.writeLong(message.getSrcId());
+    dataOutput.writeInt(message.getSrcId());
     dataOutput.writeInt((int) message.mx.countColumns());
     message.mx.asCollectable1D().supplyTo(receiveInto(dataOutput));
   }
 
   @Override
   public BackwardsSubstituteInterimMessage readMessage(DataInput dataInput) throws IOException {
-    val srcId = dataInput.readLong();
+    val srcId = dataInput.readInt();
     val dofs = dataInput.readInt();
     val ds = FACTORY.makeZero(CONTRIBUTED_ROWS, dofs);
     ds.fillMatching(dataInputAccessStore(dataInput, CONTRIBUTED_ROWS * dofs));
@@ -36,7 +36,7 @@ final class BackwardsSubstituteInterimMessageSerializer implements MessageSerial
 
   @Override
   public BackwardsSubstituteInterimMessage readMessage(BackwardsSubstituteInterimMessage message, DataInput dataInput) throws IOException {
-    val srcId = dataInput.readLong();
+    val srcId = dataInput.readInt();
     val dofs = dataInput.readInt();
     message.reattach(srcId);
     message.mx.fillMatching(dataInputAccessStore(dataInput, CONTRIBUTED_ROWS * dofs));
